enhance: Widen types for handling commodity weight and container teu count

+++ modules/cargotracker/booking/app/cargotracker-booking-commandside-app/src/main/groovy/org/klokwrk/cargotracker/booking/commandside/feature/bookingoffer/application/port/in/CargoData.groovy
@@ -60,7 +60,7 @@ class CargoData {
    */
   @Min(value = 1L, groups = [Level2])
   @NotNull(groups = [Level1])
-  Integer commodityWeightKg
+  Long commodityWeightKg
 
   /**
    * The requested storage temperature in Celsius degree for a commodity.
